[fix] log correct files
The files/dirs were not logged correctly as the paths given were not
correct.

Also, change style in how it was done, as now its just one loops.

Because of that, the progress bar is removed right now.

Also, made it so that exiftool does not run on dirs and only on files.

Signed-off-by: Jason Scheffel <contact@jasonscheffel.com>